Split test orchestration code into separate files
This is just a refactor; it was a little unmanageable to navigate around
an ~400 line file! Hopefully navigating this code will be easier with
this new organization into smaller files.
