+++ bin/test/tasks/build_fixtures.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Test::Tasks::BuildFixtures < Pallets::Task
+  def run
+    execute_rake_task('spec:fixture_builder:rebuild')
+  end
+end
