+++ bin/test/tasks/compile_java_script.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Test::Tasks::CompileJavaScript < Pallets::Task
+  def run
+    execute_system_command('bin/webpack --silent')
+  end
+end
