+++ bin/test/tasks/run_annotate.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Test::Tasks::RunAnnotate < Pallets::Task
+  def run
+    execute_system_command(<<~COMMAND)
+      bin/annotate --models --show-indexes --sort --exclude fixtures,tests && git diff --exit-code
+    COMMAND
+  end
+end
