+++ bin/test/tasks/run_brakeman.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Test::Tasks::RunBrakeman < Pallets::Task
+  def run
+    execute_system_command('bin/brakeman --quiet')
+  end
+end
