+++ bin/test/tasks/run_rubocop.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Test::Tasks::RunRubocop < Pallets::Task
+  def run
+    execute_system_command(<<~COMMAND)
+      bin/rubocop $(git ls-tree -r HEAD --name-only) --force-exclusion --format clang
+    COMMAND
+  end
+end
