+++ bin/test/tasks/run_stylelint.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Test::Tasks::RunStylelint < Pallets::Task
+  def run
+    execute_system_command('./node_modules/.bin/stylelint app/**/*.{css,scss,vue} --max-warnings 0')
+  end
+end
