+++ bin/test/tasks/setup_db.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Test::Tasks::SetupDb < Pallets::Task
+  def run
+    execute_rake_task('db:create')
+    execute_rake_task('db:environment:set', 'RAILS_ENV=test')
+    execute_rake_task('db:schema:load')
+  end
+end
